.setup/requirements.txt
# Core agent framework
langchain>=1.2.0
langchain-core>=1.2.0
langchain-openai>=1.1.0
langgraph>=1.1.0
langgraph-checkpoint>=4.0.0
langgraph-prebuilt>=1.0.0
langgraph-sdk>=0.3.0
langsmith>=0.7.0

# LLM API client
openai>=2.31.0

# Observability and tracing
langfuse>=4.1.0
opentelemetry-api>=1.41.0
opentelemetry-sdk>=1.41.0
opentelemetry-exporter-otlp-proto-http>=1.41.0

# Utilities
python-dotenv>=1.2.0
ulid-py>=1.1.0
pydantic>=2.12.0
httpx>=0.28.0

# Jupyter environment
jupyter>=1.1.0
jupyterlab>=4.5.0
ipykernel>=7.2.0
